requirements.txt
affine==2.4.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fonttools==4.58.4
geopandas==1.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
osmnx==2.0.4
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.6
Pygments==2.19.2
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.0
rasterio==1.4.3
requests==2.32.4
scipy==1.16.0
shapely==2.1.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
